Refactor: Migrate monetary values to `BigDecimal` for precision and enhance database handling
- **Core & Database:**
    - Introduced `BigDecimalConverter` to handle `BigDecimal` serialization in Room.
    - Added `BigDecimalConverter` to `BreezeDatabase` type converters and updated the database version.
    - Updated `Conta`, `ParcelaEntity`, and `MovimentacaoEntity` (implied by `BreezeDatabase` update) to use `BigDecimal` for currency values.

- **`AdicionarContaViewModel.kt`:**
    - Refactored `valorConta`, `taxaDeJurosMensal`, and `valorDasParcelas` to use `BigDecimal` for high-precision financial calculations.
    - Updated `calculaParcelasSemJuros` and `calculaParcelasComJuros` to use `BigDecimal` arithmetic and `RoundingMode.HALF_EVEN`.
    - Changed `valorConta` state to a `String` to improve input flexibility.

- **UI Components:**
    - **`DestaquesCard.kt`**: Integrated `MoneyUtils.arredondarValor` for consistent currency formatting and updated date display to use `monthValue`.
    - **`TextValue.kt`**: Temporarily reverted `value` parameter from `BigDecimal` to `Double` for compatibility with specific UI callers.
    - **`EditarValorConta.kt`**: Updated value manipulation logic to use `BigDecimal` constants.
    - **`Passo3.kt`**: Added a `modifier` parameter to the component for better layout control.

- **Models:**
    - Updated `LinhaDoTempoModel` and `ConfirmPaymentModel` to store `valor` as `BigDecimal`.

+++ app/src/main/java/com/migueldk17/breeze/data/local/entity/Conta.kt
@@ -7,6 +7,7 @@ import androidx.room.ColumnInfo
 import androidx.room.Entity
 import androidx.room.PrimaryKey
 import com.migueldk17.breeze.enums.StatusConta
+import java.math.BigDecimal
 import java.time.LocalDate
 
 
@@ -25,7 +26,7 @@ data class Conta(
     val subCategoria: String,
 
     @ColumnInfo(name = "valor")
-    val valor: Double, //Valor monetário da conta
+    val valor: BigDecimal, //Valor monetário da conta
 
     @ColumnInfo(name = "icon")
     val icon: String, //Referencia do icone de tipo BreezeIconsType
